refactor/#9-home-ui : 패키지 구조 변경
